book_bloom/bookbloom/pyproject.toml
[project]
name = "bookbloom"
version = "0.1.0"
description = "BookBloom - Books Reborn, Knowledge Renewed"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "aiosqlite>=0.19.0",
    "pydantic>=2.5.0",
    "email-validator>=2.1.0",
]
